Desktop/A_Tuo_Servizio/services/servizio_profili_professionisti/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx # Required by TestClient
